refactor: pass user object instead of permissions to CustomDataGrid. Added column parsing function in Orders and Places tables to hide columns based on user_type.
