+++ laravel/resources/js/Pages/Kids/KidContacts.jsx
@@ -7,7 +7,7 @@ import { useEffect, useState } from 'react';
 import { parse } from 'date-fns';
 import CustomDataGrid from '@/Components/CustomDataGrid';
 
-export default function KidContacts( {auth, kid, flash, permissions} ) {
+export default function KidContacts({ auth, kid, flash }) {
 
     const [openSnackbar, setOpenSnackbar] = useState(false);                // defines if snackbar shows or not
     const [snackbarMessage, setSnackbarMessage] = useState('');             // defines the message to be shown in the snackbar
@@ -191,20 +191,12 @@ export default function KidContacts( {auth, kid, flash, permissions} ) {
                                 </a>
                             </Button>
 
-                            {/* <Table
-                                data={kidPhones}
-                                columnsLabel={kidPhonesColLabels}
-                                editAction="kidPhoneNumbers.showEdit"
-                                deleteAction="kidPhoneNumbers.delete"
-                                dataId="id" // Ensure the correct field is passed for DataGrid's `id`
-                            /> */}
-
                             <CustomDataGrid 
                                 rows={kidPhones}
                                 columns={kidPhonesColumns}
                                 editAction="kidPhoneNumbers.showEdit"
                                 deleteAction="kidPhoneNumbers.delete"
-                                permissions={permissions}
+                                user={auth.user}
                             />
                         </div>
                     </div>
@@ -218,21 +210,13 @@ export default function KidContacts( {auth, kid, flash, permissions} ) {
                                     Novo Email
                                 </a>
                             </Button>
-
-                            {/* <Table
-                                data={kidEmails}
-                                columnsLabel={kidEmailsColLabels}
-                                editAction="kidEmails.showEdit"
-                                deleteAction="kidEmails.delete"
-                                dataId="id" // Ensure the correct field is passed for DataGrid's `id`
-                            /> */}
                             
                             <CustomDataGrid
                                 rows={kidEmails}
                                 columns={kidEmailColumns}
                                 editAction="kidEmails.showEdit"
                                 deleteAction="kidEmails.delete"
-                                permissions={permissions}
+                                user={auth.user}
                             />
                         </div>
                     </div>
