+++ laravel/resources/js/Pages/Orders/OrderStart.jsx
@@ -6,7 +6,7 @@ import { useState, useEffect, useContext } from "react";
 import WaypointManager from "./Partials/WaypointManager";
 import 'leaflet/dist/leaflet.css';
 
-export default function OrderStart({ auth, order, kids, otherPlaces, orderRoutes, onlyView, flash, permissions }) {
+export default function OrderStart({ auth, order, kids, otherPlaces, orderRoutes, onlyView, flash }) {
     return (
         <OrderProvider>
             <InnerOrderStart
@@ -16,14 +16,13 @@ export default function OrderStart({ auth, order, kids, otherPlaces, orderRoutes
                 otherPlaces={otherPlaces}
                 orderRoutes={orderRoutes}
                 onlyView={onlyView}
-                permissions={permissions}
             />
         </OrderProvider>
     );
 }
 
 
-function InnerOrderStart({ auth, order, kids, otherPlaces, orderRoutes, onlyView, flash, permissions }) {
+function InnerOrderStart({ auth, order, kids, otherPlaces, orderRoutes, onlyView, flash }) {
 
     const { updateWaypoints, updatePlaces, } = useContext(OrderContext);
 
@@ -249,7 +248,7 @@ function InnerOrderStart({ auth, order, kids, otherPlaces, orderRoutes, onlyView
 
                                 </Grid>
 
-                                {!onlyView && !permissions.isDriver &&
+                                {!onlyView && auth.user.user_type !== "Condutor" &&
                                     <Grid item xs={12} display="flex" justifyContent="flex-end" marginTop={2} gap="10px">
                                         <Button
                                             variant="contained"
